proc: split proc virtualization into separate files
