+++ Makefile.am
@@ -18,8 +18,11 @@ liblxcfs_la_SOURCES = bindings.c bindings.h \
 		      cgroups/cgroup.c cgroups/cgroup.h \
 		      cgroups/cgroup2_devices.c cgroups/cgroup2_devices.h \
 		      cgroups/cgroup_utils.c cgroups/cgroup_utils.h \
-		      cpuset.c \
+		      cpuset_parse.c cpuset_parse.h \
 		      memory_utils.h \
+		      proc_cpuview.c proc_cpuview.h \
+		      proc_fuse.c proc_fuse.h \
+		      proc_loadavg.c proc_loadavg.h \
 		      sysfs_fuse.c sysfs_fuse.h \
 		      utils.c utils.h
 liblxcfs_la_CFLAGS = $(AM_CFLAGS)
@@ -31,8 +34,11 @@ liblxcfstest_la_SOURCES = bindings.c bindings.h \
 			  cgroups/cgroup.c cgroups/cgroup.h \
 			  cgroups/cgroup2_devices.c cgroups/cgroup2_devices.h \
 			  cgroups/cgroup_utils.c cgroups/cgroup_utils.h \
-			  cpuset.c \
+			  cpuset_parse.c cpuset_parse.h \
 			  memory_utils.h \
+			  proc_cpuview.c proc_cpuview.h \
+			  proc_fuse.c proc_fuse.h \
+			  proc_loadavg.c proc_loadavg.h \
 			  sysfs_fuse.c sysfs_fuse.h \
 			  utils.c utils.h
 liblxcfstest_la_CFLAGS = $(AM_CFLAGS) -DRELOADTEST
@@ -62,12 +68,15 @@ lxcfs.1: lxcfs lxcfs.man.add
 endif
 
 TEST_READ: tests/test-read.c
-	$(CC) -o tests/test-read tests/test-read.c
-TEST_CPUSET: tests/cpusetrange.c cpuset.c
-	$(CC) -o tests/cpusetrange tests/cpusetrange.c cpuset.c
+	$(CC) -o tests/test-read \
+		tests/test-read.c
+TEST_CPUSET: tests/cpusetrange.c cpuset_parse.c
+	$(CC) -o tests/cpusetrange \
+		tests/cpusetrange.c \
+		cpuset_parse.c
 TEST_SYSCALLS: tests/test_syscalls.c
-	$(CC) -o tests/test_syscalls tests/test_syscalls.c
-
+	$(CC) -o tests/test_syscalls \
+		tests/test_syscalls.c
 tests: TEST_READ TEST_CPUSET TEST_SYSCALLS
 
 distclean:
