+++ sysfs_fuse.c
@@ -1,6 +1,14 @@
 /* SPDX-License-Identifier: LGPL-2.1-or-later */
 
+#ifndef _GNU_SOURCE
+#define _GNU_SOURCE
+#endif
+
+#ifndef FUSE_USE_VERSION
 #define FUSE_USE_VERSION 26
+#endif
+
+#define _FILE_OFFSET_BITS 64
 
 #define __STDC_FORMAT_MACROS
 #include <dirent.h>
