+++ utils.c
@@ -1,6 +1,18 @@
-#define __STDC_FORMAT_MACROS
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
+
+#ifndef _GNU_SOURCE
+#define _GNU_SOURCE
+#endif
+
+#ifndef FUSE_USE_VERSION
+#define FUSE_USE_VERSION 26
+#endif
+
+#define _FILE_OFFSET_BITS 64
+
 #include <errno.h>
 #include <fcntl.h>
+#include <fuse.h>
 #include <inttypes.h>
 #include <sched.h>
 #include <stdarg.h>
@@ -9,10 +21,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <sys/epoll.h>
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <unistd.h>
 
+#include "bindings.h"
 #include "config.h"
 #include "macro.h"
 #include "memory_utils.h"
@@ -142,4 +156,200 @@ void do_release_file_info(struct fuse_file_info *fi)
 	free_disarm(f->file);
 	free_disarm(f->buf);
 	free_disarm(f);
+}
+
+#define POLLIN_SET ( EPOLLIN | EPOLLHUP | EPOLLRDHUP )
+
+bool wait_for_sock(int sock, int timeout)
+{
+	struct epoll_event ev;
+	int epfd, ret, now, starttime, deltatime, saved_errno;
+
+	if ((starttime = time(NULL)) < 0)
+		return false;
+
+	if ((epfd = epoll_create(1)) < 0) {
+		lxcfs_error("%s\n", "Failed to create epoll socket: %m.");
+		return false;
+	}
+
+	ev.events = POLLIN_SET;
+	ev.data.fd = sock;
+	if (epoll_ctl(epfd, EPOLL_CTL_ADD, sock, &ev) < 0) {
+		lxcfs_error("%s\n", "Failed adding socket to epoll: %m.");
+		close(epfd);
+		return false;
+	}
+
+again:
+	if ((now = time(NULL)) < 0) {
+		close(epfd);
+		return false;
+	}
+
+	deltatime = (starttime + timeout) - now;
+	if (deltatime < 0) { // timeout
+		errno = 0;
+		close(epfd);
+		return false;
+	}
+	ret = epoll_wait(epfd, &ev, 1, 1000*deltatime + 1);
+	if (ret < 0 && errno == EINTR)
+		goto again;
+	saved_errno = errno;
+	close(epfd);
+
+	if (ret <= 0) {
+		errno = saved_errno;
+		return false;
+	}
+	return true;
+}
+
+bool recv_creds(int sock, struct ucred *cred, char *v)
+{
+	struct msghdr msg = { 0 };
+	struct iovec iov;
+	struct cmsghdr *cmsg;
+	char cmsgbuf[CMSG_SPACE(sizeof(*cred))];
+	char buf[1];
+	int ret;
+	int optval = 1;
+
+	*v = '1';
+
+	cred->pid = -1;
+	cred->uid = -1;
+	cred->gid = -1;
+
+	if (setsockopt(sock, SOL_SOCKET, SO_PASSCRED, &optval, sizeof(optval)) == -1) {
+		lxcfs_error("Failed to set passcred: %s\n", strerror(errno));
+		return false;
+	}
+	buf[0] = '1';
+	if (write(sock, buf, 1) != 1) {
+		lxcfs_error("Failed to start write on scm fd: %s\n", strerror(errno));
+		return false;
+	}
+
+	msg.msg_name = NULL;
+	msg.msg_namelen = 0;
+	msg.msg_control = cmsgbuf;
+	msg.msg_controllen = sizeof(cmsgbuf);
+
+	iov.iov_base = buf;
+	iov.iov_len = sizeof(buf);
+	msg.msg_iov = &iov;
+	msg.msg_iovlen = 1;
+
+	if (!wait_for_sock(sock, 2)) {
+		lxcfs_error("Timed out waiting for scm_cred: %s\n", strerror(errno));
+		return false;
+	}
+	ret = recvmsg(sock, &msg, MSG_DONTWAIT);
+	if (ret < 0) {
+		lxcfs_error("Failed to receive scm_cred: %s\n", strerror(errno));
+		return false;
+	}
+
+	cmsg = CMSG_FIRSTHDR(&msg);
+
+	if (cmsg && cmsg->cmsg_len == CMSG_LEN(sizeof(struct ucred)) &&
+			cmsg->cmsg_level == SOL_SOCKET &&
+			cmsg->cmsg_type == SCM_CREDENTIALS) {
+		memcpy(cred, CMSG_DATA(cmsg), sizeof(*cred));
+	}
+	*v = buf[0];
+
+	return true;
+}
+
+static int msgrecv(int sockfd, void *buf, size_t len)
+{
+	if (!wait_for_sock(sockfd, 2))
+		return -1;
+	return recv(sockfd, buf, len, MSG_DONTWAIT);
+}
+
+int send_creds(int sock, struct ucred *cred, char v, bool pingfirst)
+{
+	struct msghdr msg = { 0 };
+	struct iovec iov;
+	struct cmsghdr *cmsg;
+	char cmsgbuf[CMSG_SPACE(sizeof(*cred))];
+	char buf[1];
+	buf[0] = 'p';
+
+	if (pingfirst) {
+		if (msgrecv(sock, buf, 1) != 1) {
+			lxcfs_error("%s\n", "Error getting reply from server over socketpair.");
+			return SEND_CREDS_FAIL;
+		}
+	}
+
+	msg.msg_control = cmsgbuf;
+	msg.msg_controllen = sizeof(cmsgbuf);
+
+	cmsg = CMSG_FIRSTHDR(&msg);
+	cmsg->cmsg_len = CMSG_LEN(sizeof(struct ucred));
+	cmsg->cmsg_level = SOL_SOCKET;
+	cmsg->cmsg_type = SCM_CREDENTIALS;
+	memcpy(CMSG_DATA(cmsg), cred, sizeof(*cred));
+
+	msg.msg_name = NULL;
+	msg.msg_namelen = 0;
+
+	buf[0] = v;
+	iov.iov_base = buf;
+	iov.iov_len = sizeof(buf);
+	msg.msg_iov = &iov;
+	msg.msg_iovlen = 1;
+
+	if (sendmsg(sock, &msg, 0) < 0) {
+		lxcfs_error("Failed at sendmsg: %s.\n",strerror(errno));
+		if (errno == 3)
+			return SEND_CREDS_NOTSK;
+		return SEND_CREDS_FAIL;
+	}
+
+	return SEND_CREDS_OK;
+}
+
+int read_file_fuse(const char *path, char *buf, size_t size, struct file_info *d)
+{
+	__do_free char *line = NULL;
+	__do_fclose FILE *f = NULL;
+	size_t linelen = 0, total_len = 0;
+	char *cache = d->buf;
+	size_t cache_size = d->buflen;
+
+	f = fopen(path, "r");
+	if (!f)
+		return 0;
+
+	while (getline(&line, &linelen, f) != -1) {
+		ssize_t l = snprintf(cache, cache_size, "%s", line);
+		if (l < 0) {
+			perror("Error writing to cache");
+			return 0;
+		}
+		if (l >= cache_size) {
+			lxcfs_error("%s\n", "Internal error: truncated write to cache.");
+			return 0;
+		}
+		cache += l;
+		cache_size -= l;
+		total_len += l;
+	}
+
+	d->size = total_len;
+	if (total_len > size)
+		total_len = size;
+
+	/* read from off 0 */
+	memcpy(buf, d->buf, total_len);
+
+	if (d->size > total_len)
+		d->cached = d->size - total_len;
+	return total_len;
 }
